Move Measure tests to single file
