--- Examples/tests/Integration/MeasureSkewTMUTests.cs
@@ -1,25 +1,0 @@
-using NationalInstruments.Examples.SemiconductorTestLibrary.CodeSnippets.InstrumentAbstraction;
-using NationalInstruments.TestStand.SemiconductorModule.CodeModuleAPI;
-using Xunit;
-using static NationalInstruments.SemiconductorTestLibrary.TestStandSteps.SetupAndCleanupSteps;
-using static NationalInstruments.Tests.SemiconductorTestLibrary.Utilities.TSMContext;
-
-namespace NationalInstruments.Tests.SemiconductorTestLibrary.Integration
-{
-    [Collection("NonParallelizable")]
-    public class MeasureSkewWithSTLTest
-    {
-        private const string PinMapFileName = @"NIDigitalTMUTest.pinmap";
-        private const string DigitalProjectFileName = @"NIDigitalTMUTest.digiproj";
-        private ISemiconductorModuleContext _tsmContext = CreateTSMContext(PinMapFileName, DigitalProjectFileName);
-
-        [Fact]
-        public void InitializeNIDigital_MeasureSkewWithSTLSucceeds()
-        {
-            SetupNIDigitalPatternInstrumentation(_tsmContext);
-
-            MeasureSkewTMU.MeasureSkewWithSTL(_tsmContext);
-            CleanupInstrumentation(_tsmContext);
-        }
-    }
-}
